Split up tests and add ticket links to skipped tests
